`registry_monitor` & `uri_scheme_handler` -> `local`
